Update PULL_REQUEST_TEMPLATE.md
Consider this small rephrasing, pls. Current one confuses at start: should one try to exclude all mentioning of "S-word" in the description, or leave it only in the names, or anything else; then start to compare own description with others where it's quite common, and start to deduce why it's there, are these all are exceptions on different basis, and whether you should to continue to avoid "Solidity."

So it's just a suggestion, if this word would suit what you meant.